requirements.txt
psutil
cryptography
requests
PyYAML
networkx
numpy